% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsedFunc.R
\name{deembedy}
\alias{deembedy}
\title{deembeds m-dimensional space}
\usage{
deembedy(y, nset = 1, Td = 1, displ = F)
}
\arguments{
\item{y}{the embedded numeric matrix with n rows and m columns}

\item{nset}{integer specifying the number of data sets used to build y}

\item{Td}{integer corresponding to the embedding delay}

\item{displ}{boolean controlling the display of the matrix}
}
\value{
the averaged state vector with n+(m-1)*T rows and 1 column
}
\description{
Deembeds an m-dimensional space which has been created
by \code{embedy()}.
}
\examples{
y <- sqrt(data.mat+1)  # data.mat contains the stations in columns and time in rows
# select 8-columns
yytest <- matrix(y[!is.na(y)][1:200], ncol = 8)
y.embed <- embedy(yytest, 3 )
deembedy(y.embed)
}
