% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsedFunc.R
\name{embedy}
\alias{embedy}
\title{Embedding time-series}
\usage{
embedy(x, D, T = 1, displ = F)
}
\arguments{
\item{x}{input numeric matrix where records are in columns; n rows and m columns}

\item{D}{integer specifying the embedding dimension}

\item{T}{positive integer specifying the embedding delay. Default to 1}

\item{displ}{boolean controlling the display of the size of the embedded matrix}
}
\value{
\code{y} the embedded numeric matrix with n-D+1 rows and D*m columns
}
\description{
Embeds a set of time series c(x1, x2,..., xm) into an D-dimensional
space by taking as state vectors the consecutive sequences: (written in Matlab's style)
z(k,:) = [x1(k) x1(k+T) ... x1(k+D*T-1) x2(k) x2(k+T) ...  xm(k+D*T-1)]
for k=1 to n-D*T+1,  where n is the length of x and T is the embedding delay
}
\examples{
y <- sqrt(data.mat+1)  # data.mat contains the stations in columns and time in rows
yytest <- matrix(y[!is.na(y)][1:200], ncol = 8)
embedy(yytest, 3 )
}
