% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsedFunc.R
\name{rank_reduce}
\alias{rank_reduce}
\title{Rank reduding}
\usage{
rank_reduce(X, ncomp)
}
\arguments{
\item{X}{Numeric matrix}

\item{ncomp}{number of significant components}
}
\value{
The corresponding reduced rank version of X
}
\description{
\code{rank_reduce()} generates a low-rank version version of matrix X by computing
its SVD, and then then reconstructing X by keeping only its ncomp most important
singular values.
}
\details{
This function bypasses the (slower) full computation of the SVD by
estimating the singular vectors through diagonalisation of the covariance matrix of X
}
\examples{
rank_reduce(dataSSN, ncomp = 10)
}
