% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsedFunc.R
\name{smooth_gauss}
\alias{smooth_gauss}
\title{Gaussian smoothing}
\usage{
smooth_gauss(x, ns)
}
\arguments{
\item{x}{data matrix to smooth (smoothing is done columnwise)}

\item{ns}{smoothing width of the window}
}
\value{
\code{y.smt}	corresponding to the smoothed data
}
\description{
Smooths data in x using either a gaussian window
which has a total width given by 2*ns+1 (each side contains ns data points)
}
\examples{
y <- sqrt(data.mat+1)  # data.mat contains the stations in columns and time in rows
y_smm <- smooth_gauss(y, 8)
}
