% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsedFunc.R
\name{solar.cycle}
\alias{solar.cycle}
\title{Solar cycles on ggplots}
\usage{
solar.cycle(col = 2, size = 0.3)
}
\arguments{
\item{col}{choose color for the vertical lines}

\item{size}{choose size for the lines}
}
\value{
vertical lines plotted on ggplots showing the solar cycles
}
\description{
Funnction to plot vertical lines when added to ggplots corresponding to the solar cycles.
Dates found for the solar Cycles comes from
 \url{https://en.wikipedia.org/wiki/List_of_solar_cycles}
}
\details{
This function is useful to decrease the amount of code for each ggplots
generated as this thesis will use exclusive \code{ggplot2} for the plots.
Values of other parameters such as colours could be changed inside the function.
}
\examples{
df_uccle_vero <- data.frame(time = z1$Date[1:nrow(filledt.vero)], SSN_wnUC2 = filledt.vero$UC2, from = "Vero")

df_uccle_c <- rbind.data.frame(df_uccle_vero, data.frame(time = z1$Date[1:nrow(df_uccle_vero)],
                                                        SSN_wnUC2 = zssn[,"wnUC2"][1:nrow(df_uccle_vero)], from = "`interpolEM`"))

ggplot(df_uccle_c, aes(col = from)) +
  geom_point(aes(x = time, y = SSN_wnUC2), size = 0.15) +
  solar.cycle() +
  guides(colour = guide_legend(override.aes = list(size= 3))) +
  ggtitle("Filled missing values") +
  theme_piss()
}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
