% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross-Validation.R
\name{cvFromInterpolsvd}
\alias{cvFromInterpolsvd}
\title{cross-validation from interpolsvd_em algorithm}
\usage{
cvFromInterpolsvd(x, comp_max = 4, niter = 5, min_keep_frac = 0.1,
  seed = 123)
}
\arguments{
\item{seed}{}
}
\value{
A list with the following elements:
\describe{
  \item{\code{y.filled}}{The same dataset as y but with gaps filled}
  \item{\code{w.distSVD}}{The distribution of the weights of the SVD}
  \item{\code{errorByComp}}{Numeric vector of length \code{niter} (??)
  containing the errors associated to each iterations( or comp?)}
}
}
\description{
It aims to compare

The three tuneable (hyper)parameters are :
\describe{
\item{\code{ncomp}}
\item{\code{nsmo}}
\item{\code{nembed}}
}
}
\details{

}
\examples{

}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
