% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross-Validation.R
\name{interpol_CrossVal}
\alias{interpol_CrossVal}
\title{interpolsvd_em algorithm for cross-validation}
\usage{
interpol_CrossVal(y, nembed = 1, nsmo = 0, ncomp = 0,
  threshold1 = 1e-05, niter = 30)
}
\description{
This present a "smoother" version of \code{\link{interpolsvd_em}} to apply the
cross-validation.
}
\examples{

}
\seealso{
\code{\link{interpolsvd_em}} for information about the paramters and
the returned values which are the same.
}
\author{
Antoine Pissoort, \email{antoine.pissoort@student.uclouvain.be}
}
