% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Shiny application}
\usage{
runExample(example)
}
\arguments{
\item{example}{chooses the example you want. Run runExample() to have the list}
}
\examples{
ValUSunSSN::runExample('myapp') # and you should see the app
}
